# perf: scaffold tier1 harness with p95 latency gate