--- a/.github/workflows/perf-tier1.yml
+++ b/.github/workflows/perf-tier1.yml
@@ -1,0 +1,18 @@
+name: Perf Tier 1
+
+on:
+  workflow_dispatch:
+
+jobs:
+  perf-check:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+      - name: Setup Python
+        uses: actions/setup-python@v4
+        with:
+          python-version: 3.11
+      - name: Install deps
+        run: pip install locust
+      - name: Run perf harness
+        run: ./perf/run.sh