--- a/perf/__init__.py
+++ b/perf/__init__.py
@@ -1,0 +1,1 @@
+"""Performance test package for alfred-agent-platform-v2."""