## Summary

This PR scaffolds a performance testing harness using Locust to measure p95 latency for the RAG API endpoint.

## Changes

- Added `perf/` directory with Locust test configuration
- Created `perf-tier1.yml` workflow for manual performance testing
- Implemented p95 latency gate that fails if latency exceeds 300ms

## Test Harness Details

- **Load Pattern**: 50 concurrent users, ramping at 10 users/second
- **Test Duration**: 20 seconds
- **Endpoint**: POST `/rag/query` 
- **Failure Condition**: p95 latency > 300ms

## Notes

- This is safe to merge independently of the agent-core schema work
- The workflow is manually triggered via `workflow_dispatch`
- Expected to error cleanly with mock/null handlers until RAG endpoint is fully implemented

## How to Test

1. Go to Actions → "Perf Tier 1" workflow
2. Click "Run workflow" and select this branch
3. Confirm it runs and reports metrics (or fails if p95 > 300ms)