@locotoki 2025-05-23
## ✅ Updated Performance Harness

The performance harness has been updated with the following changes:

### API Contract Updates
- ✅ Now tests **POST /v1/query** endpoint (updated from /rag/query)
- ✅ Port updated to **8080** (from 8000)
- ✅ Validates response contains **citations** array
- ✅ Proper error handling for non-200 responses and JSON parsing

### Load Configuration
- Reduced concurrent users from 50 to 20 for more realistic testing
- Maintains 20-second test duration
- Ramp rate: 10 users/second

### Response Validation
\\\

### Ready for Integration
This harness is now ready for:
- Dashboard wiring to display p95 latency metrics
- Integration with the actual v1 API when deployed
- CI/CD pipeline inclusion for performance regression testing

The harness correctly validates the expected v1 query contract and will fail if p95 latency exceeds 300ms.